realized the interaction with mongodb to fetch book collection
