+++ app.js
@@ -12,6 +12,7 @@ var nav = [{
     Text: 'Author'
 }];
 var bookRouter = require('./src/routes/bookRoutes')(nav);
+var adminRouter = require('./src/routes/adminRoutes')(nav);
 
 app.use(express.static('public'));
 app.set('views', './src/views');
@@ -19,6 +20,7 @@ app.set('views', './src/views');
 app.set('view engine', 'ejs');
 
 app.use('/Books', bookRouter);
+app.use('/Admin', adminRouter);
 
 app.get('/', function (req, res) {
     res.render('index', {
