+++ package.json
@@ -13,7 +13,8 @@
     "ejs": "^2.3.4",
     "express": "^4.13.3",
     "express-handlebars": "^2.0.1",
-    "jade": "^1.11.0"
+    "jade": "^1.11.0",
+    "mongodb": "^2.0.46"
   },
   "devDependencies": {
     "gulp": "^3.9.0",
