+++ src/routes/bookRoutes.js
@@ -1,78 +1,55 @@
 var express = require('express');
-
 var bookRouter = express.Router();
+var mongodb = require('mongodb').MongoClient;
+var objectId = require('mongodb').ObjectID;
 
-var router = function(nav){
-    var books = [
-    {
-        title: 'War and Peace',
-        genre: 'Historical Fiction',
-        author: 'Lev Nikolayevich Tolstoy',
-        read: false
-        },
-    {
-        title: 'Les Misérables',
-        genre: 'Historical Fiction',
-        author: 'Victor Hugo',
-        read: false
-        },
-    {
-        title: 'The Time Machine',
-        genre: 'Science Fiction',
-        author: 'H. G. Wells',
-        read: false
-        },
-    {
-        title: 'A Journey into the Center of the Earth',
-        genre: 'Science Fiction',
-        author: 'Jules Verne',
-        read: false
-        },
-    {
-        title: 'The Dark World',
-        genre: 'Fantasy',
-        author: 'Henry Kuttner',
-        read: false
-        },
-    {
-        title: 'The Wind in the Willows',
-        genre: 'Fantasy',
-        author: 'Kenneth Grahame',
-        read: false
-        },
-    {
-        title: 'Life On The Mississippi',
-        genre: 'History',
-        author: 'Mark Twain',
-        read: false
-        },
-    {
-        title: 'Childhood',
-        genre: 'Biography',
-        author: 'Lev Nikolayevich Tolstoy',
-        read: false
-        }
-    ];
-    
+var router = function (nav) {
+
     bookRouter.route('/')
-    .get(function (req, res) {
-        res.render('bookListView', {
-            title: 'Books',
-            nav: nav,
-            books: books
+        .get(function (req, res) {
+            var url =
+                'mongodb://localhost:27017/libraryApp';
+
+            mongodb.connect(url, function (err, db) {
+                var collection = db.collection('books');
+
+                collection.find({}).toArray(
+                    function (err, results) {
+                        res.render('bookListView', {
+                            title: 'Books',
+                            nav: nav,
+                            books: results
+                        });
+                    }
+                );
+            });
+
         });
-    });
 
     bookRouter.route('/:id')
-    .get(function (req, res) {
-        var id = req.params.id;
-        res.render('bookView', {
-            title: 'Books',
-            nav: nav,
-            book: books[id]
+        .get(function (req, res) {
+            var id = new objectId(req.params.id);
+            var url =
+                'mongodb://localhost:27017/libraryApp';
+
+            mongodb.connect(url, function (err, db) {
+                var collection = db.collection('books');
+
+                collection.findOne({_id: id},
+                    function (err, result) {
+                        res.render('bookView', {
+                            title: 'Books',
+                            nav: nav,
+                            book: result
+                        });
+
+                    }
+                );
+
+            });
+
         });
-    });
-    
+
     return bookRouter;
-}
+};
 module.exports = router;
